requirements.txt
# ============================================================================
# Core Scientific Computing
# ============================================================================
numpy==2.3.5
scipy==1.17.0
pandas==3.0.0
scikit-learn==1.8.0
numba==0.63.1

# ============================================================================
# Deep Learning & ML Frameworks
# ============================================================================
torch==2.10.0
torchvision==0.25.0
torchtext==0.18.0
torchtune==0.6.1
torchmetrics==1.8.2
transformers==4.48.1
lightning-utilities==0.15.2
optuna==4.7.0

# ============================================================================
# ESM Models & Language Models
# ============================================================================
esm==3.2.1.post1
safetensors==0.7.0
tokenizers==0.21.4
huggingface-hub==0.36.2

# ============================================================================
# Biology & Protein Processing
# ============================================================================
biopython==1.86
biotite==1.6.0
biotraj==1.2.2

# ============================================================================
# Visualization & Plotting
# ============================================================================
matplotlib==3.10.8
seaborn==0.13.2
umap-learn==0.5.11
jupyter==1.1.1
jupyterlab==4.5.3
ipython==9.10.0

# ============================================================================
# Utilities & Infrastructure
# ============================================================================
einops==0.8.2
networkx==3.6.1
pyyaml==6.0.3
requests==2.32.5
cloudpathlib==0.23.0
fsspec==2025.10.0
tqdm==4.67.3
pillow==12.1.0
brotli==1.2.0

# ============================================================================
# Testing & Code Quality
# ============================================================================
pytest==9.0.2
pytest-cov==7.0.0
coverage==7.13.4
ruff==0.15.0
mypy==1.18.2
vulture==2.14
deptry==0.24.0
mypy==1.19.1
